apps/api/pyproject.toml
[project]
name = "myelectricaldata-api"
version = "1.1.0-dev.5"
description = "MyElectricalData API Gateway for Enedis data"
authors = [{name = "m4dm4rtig4n"}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "httpx>=0.27.2",
    "cryptography>=43.0.3",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.17",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "pdfplumber>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.2",
    "faker>=33.1.0",
    "black>=24.10.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "faker>=33.1.0",
]
